refactor(web): 将updateActor的action迁移到web相关文件夹
因为耦合了web api的逻辑
