+++ src/shared/redux/actions/actor.ts
@@ -12,7 +12,6 @@ const {
   GET_ACTOR_SUCCESS,
   SELECT_ACTOR,
   REMOVE_ACTOR_SUCCESS,
-  UPDATE_ACTOR_SUCCESS,
 } = constants;
 import config from '../../project.config';
 import { checkTemplate } from '../../utils/cache-helper';
@@ -27,11 +26,9 @@ import _isUndefined from 'lodash/isUndefined';
 import _isString from 'lodash/isString';
 import _set from 'lodash/set';
 
-import * as trpgApi from '../../api/trpg.api';
 import { TRPGAction } from '@redux/types/__all__';
-import { isBlobUrl } from '@shared/utils/string-helper';
-import { getJWTInfo } from '@portal/utils/auth';
-import { toAvatarWithBlobUrl } from '@web/utils/upload-helper';
+
+import * as trpgApi from '../../api/trpg.api';
 const api = trpgApi.getInstance();
 
 let setTemplate = function setTemplate(uuid, name, desc, avatar, info) {
@@ -300,57 +297,6 @@ let removeActor = function removeActor(uuid) {
   };
 };
 
-/**
- * 根据UUID更新角色信息
- * @param uuid 角色唯一标识
- * @param name 角色名
- * @param avatar 角色头像
- * @param desc 角色描述
- * @param info 角色信息
- */
-let updateActor = function updateActor(
-  uuid: string,
-  name: string,
-  avatar: string,
-  desc: string,
-  info: {}
-): TRPGAction {
-  return async function(dispatch, getState) {
-    dispatch(showLoading('正在更新人物卡信息，请稍后...'));
-    if (_isString(avatar) && isBlobUrl(avatar)) {
-      // 如果avatar是blob url的话, 则上传一下
-      const userUUID = getState().user.info.uuid;
-      try {
-        const { url } = await toAvatarWithBlobUrl(userUUID, avatar);
-        // 上传成功后更新属性
-        _set(info, '_avatar', url);
-        avatar = url;
-      } catch (err) {
-        dispatch(showAlert('上传头像失败'));
-        throw err;
-      }
-    }
-
-    return api.emit(
-      'actor::updateActor',
-      { uuid, name, avatar, desc, info },
-      function(data) {
-        dispatch(hideLoading());
-        dispatch(hideAlert());
-        dispatch(hideModal());
-        if (data.result) {
-          let actor = data.actor;
-          actor.avatar = config.file.getAbsolutePath(actor.avatar);
-          dispatch({ type: UPDATE_ACTOR_SUCCESS, payload: actor });
-        } else {
-          dispatch(showAlert(data.msg));
-          console.error(data.msg);
-        }
-      }
-    );
-  };
-};
-
 export {
   setTemplate,
   getTemplate,
@@ -366,5 +312,4 @@ export {
   getActor,
   selectActor,
   removeActor,
-  updateActor,
 };
