+++ src/web/containers/main/actors/ActorEdit.tsx
@@ -3,7 +3,8 @@ import { connect } from 'react-redux';
 import * as at from 'trpg-actor-template';
 import config from '@shared/project.config';
 import { showAlert } from '@shared/redux/actions/ui';
-import { createActor, updateActor } from '@shared/redux/actions/actor';
+import { createActor } from '@shared/redux/actions/actor';
+import { updateActor } from '@web/redux/action/actor';
 import ImageUploader from '@web/components/ImageUploader';
 
 import './ActorEdit.scss';
