+++ src/web/containers/main/actors/ActorList.tsx
@@ -4,9 +4,9 @@ import * as at from 'trpg-actor-template';
 import TemplateSelect from './TemplateSelect';
 import ActorCreate from '@web/components/modal/ActorCreate';
 import ActorEdit from '@web/components/modal/ActorEdit';
-import { updateActor } from '@src/shared/redux/actions/actor';
 import * as apiHelper from '@shared/utils/api-helper';
 import { showModal, showAlert } from '@shared/redux/actions/ui';
+import { updateActor } from '@web/redux/action/actor';
 import {
   selectActor,
   removeActor,
